requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.79
botocore==1.34.79
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.1
fonttools==4.51.0
h11==0.14.0
idna==3.6
importlib_resources==6.4.0
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
zipp==3.18.1
